requirements.txt
# Dependencias para el Sistema de Cierre Mensual

# Procesamiento de PDFs
pdfplumber==0.10.3
pdf2image==1.16.3
PyPDF2==3.0.1

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Excel
openpyxl==3.1.2

# Word
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
PyPDF2==3.0.1
pdf2image>=1.16.0
pdfplumber>=0.10.0
python-dateutil>=2.8.0